Automatic startup (#5)
* Ignore vscode settings file

After updating python3 on my local machine to get rid of incorrect vscode errors, my vscode settings file needed to change.

* Update run.sh

This update was tested on the VirtualBox so that when run on startup, it does not attempt to run twice or in a separate shell. Not sure quite why, but it works.

* Create autoStart.sh

This function will add the run.sh file to your autostart file, so it is run automatically in a terminal on boot.

* Update install.sh

Added autostart.sh call within the install.sh file.