Create env file (#7)
* Create createEnv.sh

This file will create a basic .env file for a user, and will give dummy keys to start.

* Created env_addition.sh

This script will take an input of the key name for the .env file and add it. I also updated the createEnv.sh to call this function, instead of doing manual echo's into the .env file.

* Update install.sh

Added createEnv.sh to be run in install.sh after the autostart feature is set up.